Add validation to store method in CarController
In the 'store' method of the 'CarController.php' file, validation rules have been added for each car attribute to ensure that the correct data is being provided. This will improve data integrity and prevent potential errors or issues within the application. Also, the validated data is now used to create a new 'Car' object.