Add deletion cascade and UI updates for invoices and cars
Enhancements were made to invoices and cars management by adding conditions for display of the delete button. This ensures that the button only shows if an invoice or car has not been trashed. Also, code for cascade deletion was added to the Reservation model, so, when a reservation is deleted, all its associated invoices get automatically deleted. Furthermore, a small language correction was made from English to French.